streamlit-rag-app/requirements.txt
streamlit
openai>=1.0.0
pdfplumber
pymupdf
tiktoken
qdrant-client
azure-storage-blob
